+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/Mage.java
@@ -794,6 +794,8 @@ public interface Mage extends CostReducer, CooldownReducer {
     Set<String> getModifierKeys();
     @Nullable
     MageModifier getModifier(String key);
+    @Nonnull
+    Collection<MageModifier> getModifiers();
     boolean isBypassEnabled();
     void setBypassEnabled(boolean enable);
     @Nullable